requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Task scheduling
apscheduler==3.10.4

# LLM integrations
openai==1.10.0
anthropic==0.18.1
langchain==0.1.6
langchain-community==0.0.19
ollama==0.1.6

# Google Calendar
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Notion
notion-client==2.2.1

# Email
imapclient==3.0.1
email-validator==2.1.0

# Utilities
httpx==0.26.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1
rich==13.7.0
typer==0.9.0
